Favorites and more (read desc)
This update adds Favorites, alongside this a few changes to some UI was made mostly in relation to the new favorites feature.

Because of an issue, some mods had to have their id changed to remove spaces and special characters. This includes Slarmoo's Box and JummBox 11Edo.

The website now also saves which tab you were viewing. Say if you were looking at your favorites, next time you visit the website it'll automatically be set to the tab you were last on. I'll set this to be optional when I get to adding preferences.