Storing interactions with mods (See Desc)
The plan is to add a sorting mode called 'Interactions' that sorts mods based on how many times you interact with them. With the most interacted mods being first in the list and the last being ones you've interacted with the least.

In cases where the mod has yet to be interacted with, it'll sort them by date of release.